requirements.dev.txt
gitpython
jsonschema
jsonpickle
mkdocs
multiprocessing_logging
pychalk
pygithub
commentjson
pytablewriter
pytest-cov
pytest-timeout
pyyaml
requests==2.24.0
terminaltables
webpreview
yq
importlib-metadata>=3.10
mkdocs-material
mdx_truly_sane_lists
beautifulsoup4
giturlparse
json-schema-for-humans
pyyaml
json5
